Use non-strict mode when parsing global config
The HyperShift operator uses the latest version of the OpenShift API to
serialize global config in the HCP (for compatibility with older CPOs).
The issue is that if there's a new field in the latest API that the
older CPO does not understand, we currently produce an error decoding
the YAML because the YAML serializer we use is using strict mode.

This commit switches to a YAML serializer that does have strict mode for
parsing of the global configuration. Making it possible for older CPOs
to parse YAML from the latest HyperShift operator.